Move observable into decorators module
